# openapi.json
{"openapi":"3.0.1","info":{"title":"Recipe Finder API","description":"Find recipes based on available ingredients and recipe filters.","contact":{"name":"Recipe Finder Team"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8285/api","description":"Generated server url"}],"tags":[{"name":"Recipes","description":"Recipe search and retrieval endpoints"}],"paths":{"/recipes/search":{"post":{"tags":["Recipes"],"summary":"Search recipes by available ingredients","description":"Returns recipes that can be made with all supplied base ingredients and optional filters.","operationId":"searchRecipes","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"searchRequest":{"$ref":"#/components/schemas/RecipeSearchRequest"},"pageable":{"$ref":"#/components/schemas/Pageable"}}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponseRecipeDTO"}}}}}}},"/recipes":{"get":{"tags":["Recipes"],"summary":"Get all recipes","operationId":"getAllRecipes","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponseRecipeDTO"}}}}}}},"/recipes/{id}":{"get":{"tags":["Recipes"],"summary":"Get a recipe by ID","operationId":"getRecipeById","parameters":[{"name":"id","in":"path","description":"Recipe database ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RecipeDTO"}}}}}}},"/recipes/time/{maxTime}":{"get":{"tags":["Recipes"],"summary":"Get recipes by maximum cook time","operationId":"getByMaxCookTime","parameters":[{"name":"maxTime","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponseRecipeDTO"}}}}}}}},"components":{"schemas":{"RecipeSearchRequest":{"type":"object","properties":{"availableBaseIngredients":{"maxItems":100,"minItems":0,"type":"array","items":{"maxLength":100,"minLength":0,"type":"string"}},"maxCookTime":{"minimum":0,"type":"integer","format":"int32"},"searchByTitle":{"type":"string"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"IngredientDetail":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"baseIngredient":{"type":"string"},"quantityWithUnit":{"type":"string"}}},"PageResponseRecipeDTO":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/RecipeDTO"}},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"}}},"RecipeDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"cookTimeMinutes":{"type":"integer","format":"int32"},"prepTimeMinutes":{"type":"integer","format":"int32"},"servings":{"type":"integer","format":"int32"},"instructions":{"type":"string"},"author":{"type":"string"},"ratings":{"type":"number","format":"double"},"imageUrl":{"type":"string"},"ingredients":{"type":"array","items":{"$ref":"#/components/schemas/IngredientDetail"}}}}}}}